+++ App/Viewer/Resource/Translations/RabbitRemoteControlApp_zh_TW.ts
@@ -4,42 +4,42 @@
 <context>
     <name>CFavoriteView</name>
     <message>
-        <location filename="../../FavoriteView.cpp" line="200"/>
+        <location filename="../../FavoriteView.cpp" line="209"/>
         <source>Connect</source>
         <translation>連接</translation>
     </message>
     <message>
-        <location filename="../../FavoriteView.cpp" line="201"/>
+        <location filename="../../FavoriteView.cpp" line="210"/>
         <source>Open settings and connect</source>
         <translation>打開設置再連接</translation>
     </message>
     <message>
-        <location filename="../../FavoriteView.cpp" line="204"/>
+        <location filename="../../FavoriteView.cpp" line="213"/>
         <source>New group</source>
         <translation>新建組</translation>
     </message>
     <message>
-        <location filename="../../FavoriteView.cpp" line="206"/>
+        <location filename="../../FavoriteView.cpp" line="215"/>
         <source>Delete</source>
         <translation>刪除</translation>
     </message>
     <message>
-        <location filename="../../FavoriteView.cpp" line="240"/>
+        <location filename="../../FavoriteView.cpp" line="249"/>
         <source>Input</source>
         <translation>輸入</translation>
     </message>
     <message>
-        <location filename="../../FavoriteView.cpp" line="240"/>
+        <location filename="../../FavoriteView.cpp" line="249"/>
         <source>Input group name</source>
         <translation>輸入組名</translation>
     </message>
     <message>
-        <location filename="../../FavoriteView.cpp" line="245"/>
+        <location filename="../../FavoriteView.cpp" line="254"/>
         <source>Error</source>
         <translation>錯誤</translation>
     </message>
     <message>
-        <location filename="../../FavoriteView.cpp" line="245"/>
+        <location filename="../../FavoriteView.cpp" line="254"/>
         <source>The group [%1] is existed</source>
         <translation>組 [%1] 已經存在</translation>
     </message>
@@ -135,22 +135,22 @@
         <translation>普通</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="282"/>
+        <location filename="../../ParameterDlgSettings.ui" line="258"/>
         <source>Shot screen</source>
         <translation>截屏</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="307"/>
+        <location filename="../../ParameterDlgSettings.ui" line="283"/>
         <source>No action</source>
         <translation>無動作</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="323"/>
+        <location filename="../../ParameterDlgSettings.ui" line="299"/>
         <source>Open folder</source>
         <translation>打開文件夾</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="336"/>
+        <location filename="../../ParameterDlgSettings.ui" line="312"/>
         <source>Open file</source>
         <translation>打開文件</translation>
     </message>
@@ -176,7 +176,7 @@
     </message>
     <message>
         <location filename="../../ParameterDlgSettings.ui" line="88"/>
-        <location filename="../../ParameterDlgSettings.ui" line="251"/>
+        <location filename="../../ParameterDlgSettings.ui" line="328"/>
         <source>No</source>
         <translation>無</translation>
     </message>
@@ -206,72 +206,77 @@
         <translation>主窗口</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="170"/>
+        <location filename="../../ParameterDlgSettings.ui" line="238"/>
         <source>Main window receiver short cut key</source>
         <translation>主窗口接收快捷鍵</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="177"/>
+        <location filename="../../ParameterDlgSettings.ui" line="170"/>
         <source>Save main window status</source>
         <translation>保存主窗口狀態</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="184"/>
+        <location filename="../../ParameterDlgSettings.ui" line="177"/>
         <source>Tab position</source>
         <translation>標簽位置</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="196"/>
+        <location filename="../../ParameterDlgSettings.ui" line="189"/>
         <source>North</source>
         <translation>上</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="209"/>
+        <location filename="../../ParameterDlgSettings.ui" line="202"/>
         <source>South</source>
         <translation>下</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="222"/>
+        <location filename="../../ParameterDlgSettings.ui" line="215"/>
         <source>West</source>
         <translation>左</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="235"/>
+        <location filename="../../ParameterDlgSettings.ui" line="228"/>
         <source>East</source>
         <translation>右</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="245"/>
+        <location filename="../../ParameterDlgSettings.ui" line="322"/>
         <source>System tray icon context menu</source>
         <translation>系統托盤圖標右鍵菜單</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="258"/>
+        <location filename="../../ParameterDlgSettings.ui" line="335"/>
         <source>Remote</source>
         <translation>遠程菜單</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="265"/>
+        <location filename="../../ParameterDlgSettings.ui" line="342"/>
         <source>Recent open</source>
         <translation>最近打開菜單</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="272"/>
+        <location filename="../../ParameterDlgSettings.ui" line="349"/>
         <source>Favorite</source>
         <translation>收藏夾</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="294"/>
+        <location filename="../../ParameterDlgSettings.ui" line="270"/>
         <source>Select shot remote desktop, otherwise shot window</source>
         <translation>選擇截取遠程桌面，否則截取窗口</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="378"/>
+        <location filename="../../ParameterDlgSettings.ui" line="359"/>
+        <source>Resume the connections when it was last closed at startup</source>
+        <translation>啟動時，恢復上次關閉時的連接</translation>
+    </message>
+    <message>
+        <location filename="../../ParameterDlgSettings.ui" line="385"/>
         <source>Ok(&amp;O)</source>
         <translation>確定(&amp;O)</translation>
     </message>
     <message>
-        <location filename="../../ParameterDlgSettings.ui" line="385"/>
+        <location filename="../../ParameterDlgSettings.ui" line="392"/>
         <source>No(&amp;N)</source>
         <translation>取消(&amp;N)</translation>
     </message>
@@ -333,7 +338,7 @@
     </message>
     <message>
         <location filename="../../mainwindow.ui" line="227"/>
-        <location filename="../../mainwindow.cpp" line="271"/>
+        <location filename="../../mainwindow.cpp" line="273"/>
         <source>Full screen(&amp;F)</source>
         <translation>全屏(&amp;F)</translation>
     </message>
@@ -582,9 +587,9 @@
         <location filename="../../mainwindow.ui" line="233"/>
         <location filename="../../mainwindow.ui" line="236"/>
         <location filename="../../mainwindow.ui" line="239"/>
-        <location filename="../../mainwindow.cpp" line="272"/>
-        <location filename="../../mainwindow.cpp" line="273"/>
         <location filename="../../mainwindow.cpp" line="274"/>
+        <location filename="../../mainwindow.cpp" line="275"/>
+        <location filename="../../mainwindow.cpp" line="276"/>
         <source>Full screen</source>
         <translation>全屏</translation>
     </message>
@@ -594,45 +599,45 @@
         <translation>收藏夾</translation>
     </message>
     <message>
-        <location filename="../../mainwindow.cpp" line="301"/>
+        <location filename="../../mainwindow.cpp" line="303"/>
         <source>Exit full screen(&amp;E)</source>
         <translation>退出全屏(&amp;E)</translation>
     </message>
     <message>
-        <location filename="../../mainwindow.cpp" line="302"/>
-        <location filename="../../mainwindow.cpp" line="303"/>
         <location filename="../../mainwindow.cpp" line="304"/>
+        <location filename="../../mainwindow.cpp" line="305"/>
+        <location filename="../../mainwindow.cpp" line="306"/>
         <source>Exit full screen</source>
         <translation>退出全屏</translation>
     </message>
     <message>
-        <location filename="../../mainwindow.cpp" line="508"/>
-        <location filename="../../mainwindow.cpp" line="526"/>
+        <location filename="../../mainwindow.cpp" line="510"/>
+        <location filename="../../mainwindow.cpp" line="528"/>
         <source>Load file fail: </source>
         <translation>加載文件失敗：</translation>
     </message>
     <message>
-        <location filename="../../mainwindow.cpp" line="518"/>
+        <location filename="../../mainwindow.cpp" line="520"/>
         <source>Open rabbit remote control file</source>
         <translation>打開玉兔遠程控製文件</translation>
     </message>
     <message>
-        <location filename="../../mainwindow.cpp" line="520"/>
+        <location filename="../../mainwindow.cpp" line="522"/>
         <source>Rabbit remote control Files (*.rrc);;All files(*.*)</source>
         <translation>玉兔遠程控製文件(*.rrc);;所有文件(*.*)</translation>
     </message>
     <message>
-        <location filename="../../mainwindow.cpp" line="608"/>
+        <location filename="../../mainwindow.cpp" line="610"/>
         <source>Connecting to </source>
         <translation>正在連接 </translation>
     </message>
     <message>
-        <location filename="../../mainwindow.cpp" line="634"/>
+        <location filename="../../mainwindow.cpp" line="636"/>
         <source>Connected to </source>
         <translation>連接到 </translation>
     </message>
     <message>
-        <location filename="../../mainwindow.cpp" line="773"/>
+        <location filename="../../mainwindow.cpp" line="817"/>
         <source>Save screenslot to </source>
         <translation>保存截屏到 </translation>
     </message>